app/lambda/requirements.txt
typedb-driver